# api-doc/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "version": "1.0.0",
    "title": "neit Api Documentation",
    "description": "neit Api Documentation",
    "license": {
      "name": "Specify your license",
      "url": "https://yourlicense.com"
    },
    "x-logo": {
      "url": "logo.png",
      "description": "neit Logo",
      "altText": "neit Logo",
      "backgroundColor": "#483D8B"
    },
    "contact": {
      "name": "neit API Support",
      "url": "https://neit.com/info/support",
      "email": "support@neit.com"
    },
    "termsOfService": "https://neit.com/info/terms_and_conditions"
  },
  "servers": [
    {
      "url": "http://localhost:8060/api",
      "description": "Api base url"
    }
  ],
  "paths": {
...
    "/soporte": {
      "get": {
        "summary": "List Soporte",
        "description": "Soporte",
        "operationId": "ListSoporte",
        
        "tags": [
          "Soporte"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "description": "Filter table records by search value",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/SearchQuery"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Maximum number of records to select. Default 20",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/LimitQuery"
            }
          },
          {
            "name": "page",
            "in": "query",
            "description": "Current page record",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/PageNumberQuery"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ListSoporte"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/soporte/view/{id}": {
      "get": {
        "summary": "View Soporte",
        "description": "Soporte Details",
        "operationId": "ViewSoporte",
        
        "tags": [
          "Soporte"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewSoporte"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/soporte/add": {
      "post": {
        "summary": "Add Soporte",
        "description": "Add New Soporte",
        "operationId": "AddSoporte",
        
        "tags": [
          "Soporte"
        ],
        
        "requestBody": {
          "description": "Add Soporte",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AddSoporte"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ViewSoporte"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/soporte/edit/{id}": {
      "put": {
        "summary": "Edit Soporte",
        "description": "Edit Soporte",
        "operationId": "EditSoporte",
        
        "tags": [
          "Soporte"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "requestBody": {
          "description": "Edit Soporte",
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/EditSoporte"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EditSoporte"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/soporte/delete/{id}": {
      "delete": {
        "summary": "Delete Soporte",
        "description": "Soporte",
        "operationId": "DeleteSoporte",
        
        "tags": [
          "Soporte"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Select record by ID",
            "required": true,
            "type": "string",
            "schema": {
              "$ref": "#/components/schemas/RecordID"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Request Completed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "default": {
            "description": "Unable to complete request",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ListSoporte": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "nombre": {
            "description": "Nombre",
            "type": "string"
          }
        }
      },
      "ViewSoporte": {
        "type": "object",
        "properties": {
          "id": {
            "description": "Id",
            "type": "string"
          },
          "nombre": {
            "description": "Nombre",
            "type": "string"
          }
        }
      },
      "AddSoporte": {
        "type": "object",
        "properties": {
          "nombre": {
            "description": "Nombre",
            "type": "string"
          }
        }
      },
      "EditSoporte": {
        "type": "object",
        "properties": {
          "nombre": {
            "description": "Nombre",
            "type": "string"
          }
        }
      },
...
      "SearchQuery": {
        "type": "string",
        "properties": {
          "search": {
            "description": "Search value",
            "type": "string"
          }
        }
      },
...
      "LimitQuery": {
        "type": "string",
        "properties": {
          "limit": {
            "description": "Maximum number of records to select",
            "type": "integer"
          }
        }
      },
      "PageNumberQuery": {
        "type": "string",
        "properties": {
          "pagenumber": {
            "description": "Current page number",
            "type": "integer"
          }
        }
      },
      "RecordID": {
        "type": "string",
        "properties": {
          "id": {
            "description": "Record ID",
            "type": "string"
          }
        }
      },
      "Error": {
        "type": "object",
        "properties": {
          "code": {
            "description": "Http Status Code",
            "type": "integer"
          },
          "message": {
            "description": "Unable to complete request",
            "type": "string"
          }
        }
      }
    }
  }
}